\name{Maaslin-package}
\alias{Maaslin-package}
\docType{package}
\title{
MaAsLin: Multivariate statistical framework that finds associations between clinical metadata and microbial community abundance or function.
}
\description{
MaAsLin is a multivariate statistical framework that finds associations between clinical metadata and microbial community abundance or function. 
 The clinical metadata can be of any type continuous (for example age and weight), boolean (sex, stool/biopsy), or discrete/factor (cohort groupings and phenotypes). 
 MaAsLin is best used in the case when you are associating many metadata with microbial measurements.
}
\details{
\tabular{ll}{
Package: \tab Maaslin\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-01-01\cr
License: \tab MIT\cr
}
}
\author{
Timothy Tickle<ttickle@hsph.harvard.edu>,\cr
Curtis Huttenhower <chuttenh@hsph.harvard.edu>\cr
Maintainers: Ayshwarya Subramanian<subraman@broadinstitute.org>,\cr 
Lauren McIver<lauren.j.mciver@gmail.com>,\cr
George Weingart<george.weingart@gmail.com>
}
 
